Teardown glass_box_umap.py
- Also remove callbacks.py (use default checkpointing)
