+++ src/glass_box_umap/core.py
@@ -1,83 +1,14 @@
 from __future__ import annotations
 import os
-import random
 from typing import Any
 
 import numpy as np
 import pandas as pd
 import torch
-from torch import nn
 
+from .models import DeepReLUNet
 from .parametric_umap import ParametricUMAP
-from .utils import get_accelerator
-
-
-class LayerNormDetached(nn.Module):
-    """
-    A LayerNorm implementation where the variance calculation is detached from the
-    computation graph during evaluation, potentially stabilizing training.
-    """
-
-    def __init__(self, emb_dim: int):
-        super().__init__()
-        self.scale = nn.Parameter(torch.ones(emb_dim))
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward pass for LayerNormDetached."""
-        mean = x.mean(dim=-1, keepdim=True)
-        # Detach variance calculation during evaluation
-        if not self.training:
-            var = x.clone().detach().var(dim=-1, keepdim=True, unbiased=False)
-        else:
-            var = x.var(dim=-1, keepdim=True, unbiased=False)
-
-        norm_x = (x - mean) / torch.sqrt(var + 1e-12)  # Added epsilon for stability
-        return self.scale * norm_x
-
-
-class DeepReLUNet(nn.Module):
-    """
-    A deep neural network using PReLU activation and LayerNormDetached.
-    """
-
-    def __init__(self, input_size: int = 50, hidden_size: int = 256, output_size: int = 2):
-        super().__init__()
-        self.model = nn.Sequential(
-            nn.Linear(input_size, hidden_size, bias=False),
-            nn.PReLU(),
-            LayerNormDetached(hidden_size),
-            nn.Linear(hidden_size, hidden_size, bias=False),
-            nn.PReLU(),
-            LayerNormDetached(hidden_size),
-            nn.Linear(hidden_size, hidden_size, bias=False),
-            nn.PReLU(),
-            LayerNormDetached(hidden_size),
-            nn.Linear(hidden_size, hidden_size, bias=False),
-            nn.PReLU(),
-            LayerNormDetached(hidden_size),
-            nn.Linear(hidden_size, hidden_size, bias=False),
-            nn.PReLU(),
-            nn.Linear(hidden_size, output_size, bias=False),
-        )
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """Forward pass through the PReLU network."""
-        return self.model(x)
-
-
-def set_global_seeds(seed: int):
-    """Sets global seeds for reproducibility."""
-    np.random.seed(seed)
-    random.seed(seed)
-    torch.manual_seed(seed)
-    if torch.cuda.is_available():
-        torch.cuda.manual_seed_all(seed)
-    # Optional: For the Pytorch Lightning trainer
-    # pl.seed_everything(seed)
-
-    # You might also want deterministic algorithms
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
+from .utils import get_accelerator, set_global_seeds
 
 
 class GlassBoxUMAP:
