+++ src/glass_box_umap/parametric_umap/__init__.py
@@ -1,4 +1,4 @@
-from .api import (
+from .core import (
     ParametricUMAP,
     load_pumap,
 )
