+++ src/glass_box_umap/parametric_umap/lightning/__init__.py
@@ -1,9 +1,7 @@
-from .callbacks import PeriodicCheckpoint
 from .data import UMAPDataModule
 from .system import UMAPLightningModule
 
 __all__ = [
-    "PeriodicCheckpoint",
     "UMAPDataModule",
     "UMAPLightningModule",
 ]
